Move range_lexer -> lexer, range_parser -> parser
